feat: increment user assertion's revision
The user acounts on Ubuntu Core devices can be created using User
Assertions. Much like other assertion types these are versioned using
the revision property. The revision of the new assertion must be greater
than the revision of the assertion on the system and if it is not then
such a new assertion is rejected.

Up until now the revision of the User Assertion, in Serial Vault, has
been hardcoded to the value 1. As a consequence it is not possible to
regenerate a user assertion for the same credentials that would work on
a system because of the revision field being the same.

This commit makes Serial Vault track the user assertion's revision in a
database. It uses the model id and user's email as a key.